restructuring of the project for better comprehensibility
