+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/MetadataImporter.java
@@ -33,6 +33,8 @@ import eu.clarin.cmdi.vlo.StringUtils;
 import eu.clarin.cmdi.vlo.config.DataRoot;
 import eu.clarin.cmdi.vlo.config.VloConfig;
 import eu.clarin.cmdi.vlo.importer.ResourceStructureGraph.CmdiVertex;
+import eu.clarin.cmdi.vlo.importer.processor.*;
+
 import java.time.LocalDate;
 import java.time.ZoneId;
 import static java.time.temporal.ChronoUnit.DAYS;
