+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/AvailabilityPostProcessor.java
@@ -1,6 +1,8 @@
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import eu.clarin.cmdi.vlo.config.VloConfig;
+import eu.clarin.cmdi.vlo.importer.CMDIData;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
