+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/IdPostProcessor.java
@@ -1,6 +1,8 @@
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import eu.clarin.cmdi.vlo.StringUtils;
+import eu.clarin.cmdi.vlo.importer.CMDIData;
+
 import java.util.ArrayList;
 import java.util.List;
