+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/LicenseTypePostProcessor.java
@@ -1,10 +1,12 @@
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
 import eu.clarin.cmdi.vlo.FacetConstants;
 import eu.clarin.cmdi.vlo.config.VloConfig;
+import eu.clarin.cmdi.vlo.importer.CMDIData;
+
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
