+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/NamePostProcessor.java
@@ -14,12 +14,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import eu.clarin.cmdi.vlo.importer.CMDIData;
 
 /**
  *
