+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/PostProcessor.java
@@ -1,6 +1,8 @@
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import java.util.List;
+
+import eu.clarin.cmdi.vlo.importer.CMDIData;
 
 /**
  * Defines the interface for a postprocessor.
