+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/processor/PostProcessorsWithVocabularyMap.java
@@ -1,7 +1,8 @@
-package eu.clarin.cmdi.vlo.importer;
+package eu.clarin.cmdi.vlo.importer.processor;
 
 import eu.clarin.cmdi.vlo.MappingDefinitionResolver;
 import eu.clarin.cmdi.vlo.config.VloConfig;
+
 import java.io.InputStream;
 import java.util.Map;
