+++ vlo-importer/src/test/java/eu/clarin/cmdi/vlo/importer/CMDIDataProcessorTest.java
@@ -1,6 +1,8 @@
 package eu.clarin.cmdi.vlo.importer;
 
 import eu.clarin.cmdi.vlo.FacetConstants;
+import eu.clarin.cmdi.vlo.importer.processor.CMDIDataProcessor;
+
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
