+++ vlo-importer/src/test/java/eu/clarin/cmdi/vlo/importer/OrganizationPostProcessorTest.java
@@ -4,6 +4,9 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
+import eu.clarin.cmdi.vlo.importer.processor.OrganisationPostProcessor;
+import eu.clarin.cmdi.vlo.importer.processor.PostProcessor;
+
 public class OrganizationPostProcessorTest extends ImporterTestcase {
